Add explict relocation spec
