+++ spec/services/notifications/aggregated_journal_service_integration_spec.rb
@@ -1,0 +1,53 @@
+#-- encoding: UTF-8
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) 2012-2021 the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+require 'spec_helper'
+require_relative './mentioned_journals_shared'
+
+describe Notifications::AggregatedJournalService, 'integration', type: :model do
+  include_context 'with a mentioned work package being updated again'
+
+  it 'will relocate the notification to the newer journal' do
+    trigger_comment!
+
+    expect(mentioned_notification).to be_present
+    journal = mentioned_notification.journal
+
+    update_assignee!
+
+    expect(mentioned_notification.reload).to be_present
+    expect(mentioned_notification.journal).not_to eq journal
+    expect { journal.reload }.to raise_error(ActiveRecord::RecordNotFound)
+
+    # Expect only one notification to be present
+    expect(Notification.where(recipient: recipient, reason: :mentioned).count)
+      .to eq 1
+  end
+end
