+++ spec/services/notifications/mail_service_spec.rb
@@ -31,6 +31,8 @@
 require 'spec_helper'
 
 describe Notifications::MailService, type: :model do
+  require_relative './mentioned_journals_shared'
+
   subject(:call) { instance.call }
 
   let(:recipient) do
