+++ spec/services/notifications/mentioned_journals_shared.rb
@@ -1,0 +1,103 @@
+#-- encoding: UTF-8
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) 2012-2021 the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+require 'spec_helper'
+
+shared_context 'with a mentioned work package being updated again' do
+  let(:project) { FactoryBot.create :project }
+
+  let(:work_package) do
+    FactoryBot.create(:work_package, project: project).tap do |wp|
+      # Clear the initial journal job
+      wp.save!
+      clear_enqueued_jobs
+    end
+  end
+
+  let(:role) do
+    FactoryBot.create :role, permissions: %w[view_work_packages edit_work_packages]
+  end
+
+  let(:recipient) do
+    FactoryBot.create :user,
+                      preferences: {
+                        immediate_reminders: {
+                          mentioned: true
+                        }
+                      },
+                      notification_settings: [
+                        FactoryBot.build(:notification_setting,
+                                         mentioned: true,
+                                         involved: true)
+                      ],
+                      member_in_project: project,
+                      member_through_role: role
+  end
+  let(:actor) do
+    FactoryBot.create :user,
+                      member_in_project: project,
+                      member_through_role: role
+  end
+
+  let(:comment) do
+    <<~NOTE
+      Hello <mention class="mention" data-type="user" data-id="#{recipient.id}" data-text="@#{recipient.name}">@#{recipient.name}</mention>
+    NOTE
+  end
+
+  let(:mentioned_notification) do
+    Notification.find_by(recipient: recipient, journal: work_package.journals.last, reason: :mentioned)
+  end
+
+  let(:assigned_notification) do
+    Notification.find_by(recipient: recipient, journal: work_package.journals.last, reason: :assigned)
+  end
+
+  def trigger_comment!
+    User.execute_as(actor) do
+      work_package.journal_notes = comment
+      work_package.save!
+    end
+
+    perform_enqueued_jobs
+    work_package.reload
+  end
+
+  def update_assignee!
+    clear_enqueued_jobs
+
+    User.execute_as(actor) do
+      work_package.assigned_to = recipient
+      work_package.save!
+    end
+
+    perform_enqueued_jobs
+    work_package.reload
+  end
+end
